Wrote own log in alert dialog
Source: http://developer.android.com/guide/topics/ui/dialogs.html
Has two buttons: The top one brings up the log in pop-up from a method,
the other brings up the log in pop-up from a class. The difference
between the two is that the latter can be used to pass events back to
the dialogs host.